Implemented new "Hobo Power:" float modifier.  Added it to the relevant items, outfit, and effect.
Implemented new "Hobo Powered" boolean modifier.  On an item with this flag set, any numeric modifiers over 200 will be interpreted as a conversion factor (divided by 1000) from total Hobo Power to the specific modifier.  Smaller numbers are interpreted as fixed modifiers, as usual - this split is needed to handle the Imaginary Hamster, which has a mixture of fixed and variable enchantments.  Added this flag to the relevant Hodgman off-hands.


git-svn-id: svn://svn.code.sf.net/p/kolmafia/code@6339 b29ace70-8910-0410-8dcc-aa2fc6433167