Do not be so aggressive in resetting "last item used". Do so only when a new "use item" request is detected. This should fix restoring your old familiar after using the black market map automatically makes and uses the blackbird.
Fix missing use links for acquiring an abridged dictionary (to the untinker) or
a dictionary (to the chasm).

If you find any of (spooky temple map, spooky sapling, spooky fertilizer), give
a use link to the map if and only if you now have all three items.

If you find any of (black market map, broken wings, sunken eyes), give a use
link to the map if and only if you now have all three items.


git-svn-id: svn://svn.code.sf.net/p/kolmafia/code@6606 b29ace70-8910-0410-8dcc-aa2fc6433167