+ML no longer has an effect on combat rate.
Phase 1 of modifiers overhaul:  Modifiers are now recalculated every time the
selected adventure location changes.  Modifiers are processed for the
location name (prefixed with "loc:") and zone (prefixed with "zone:").  loc()
and zone() functions added to the modifier expression language to allow any
numeric value to be affected by the location.

Added numeric modifiers "Meat Drop Penalty" (needed to express how meat drops
work undersea) and "Familiar Weight (hidden)" (for undersea effects that
don't change the displayed weight).

Added boolean modifier "Variable".  This will be true for any object which is
a location or zone modifier, which contains any expressions, or which
otherwise has any special-case calculation of values.  Scripters take note -
you shouldn't cache any modifiers for objects with this flag set, as they are
not constant.

Added some additional numeric modifiers, and additional processing of the
equipped familiar item, for use in improved handing of familiar effects. 
None of this is actually functional yet.


git-svn-id: svn://svn.code.sf.net/p/kolmafia/code@6771 b29ace70-8910-0410-8dcc-aa2fc6433167