Puffer Fish is really Pufferfish.
Monster starting with "The" (like "The Man" or "The Astronomer") should retain
the initial "The" for encounter logging.

Fix tavern square adventuring: rather than having a single function which is
called twice, once before running the request and once after, which detects
which context it is called in and does different things, provide two separate
functions, each of which does exactly what is needed in the appropriate context.


git-svn-id: svn://svn.code.sf.net/p/kolmafia/code@6812 b29ace70-8910-0410-8dcc-aa2fc6433167