Don't register semirare info if the encounter was the result of using a Spooky Putty sheet.
Extends the recent locations feature to compact mode as well as regular. 
Note that this will look VERY strange if you switch modes during a session,
since the links are in different formats depending on the mode in which they
were first encountered.  It will look better on your next session, or if you
visit enough different places to push the old mode's locations off of the
list.  I'm assuming that nobody switches between modes on a regular basis...


git-svn-id: svn://svn.code.sf.net/p/kolmafia/code@6919 b29ace70-8910-0410-8dcc-aa2fc6433167