Fix calculation of the date of Easter (had some missing parentheses).
Added the fact that certain item uses may consume adventures - any item that
can lead to a combat, and llama gongs when the choiceadv settings select
Roachform.  It is now safe to do things like "use * drum machine" without
missing counter expirations (although this doesn't do anything to help with
item usage from the relay browser).

If a notice is given for a counter expiring more than two turns in the future
(this can happen when using a gym, for example), don't delete the counter -
let the warning be given again when the player reaches that point again.

If multiple counters expire on the same request, display the others in the
CLI - so that there's at least some evidence left of what happened to them.


git-svn-id: svn://svn.code.sf.net/p/kolmafia/code@7045 b29ace70-8910-0410-8dcc-aa2fc6433167