A change to how the 3rd field in tradeitems.txt is handled: instead of a numeric usage type code, it is now human-readable text.  There can be one primary usage ("eat", "drink", "usable", "multiple", "hat", "weapon", etc.), plus zero or more secondary usages, separated by commas.
The possible secondary usages are:
"usable", "reusable" - for items that have a [use], [fold], or similar link,
in addition to their primary [equip] or other function.  Such items no longer
have to be hard-coded in mafia.
"combat", "combat reusable" - for items usable in combat, which is now
entirely independent of other usages.
"single", "solo" - for combat items only, specifies that the item should not
be used twice, or that the second slot should not be used at all, if the
player has Funkslinking.
"bounty" - this was previously hard-coded.
"curse" - for items that can be used on other players.  We're not currently
handling such items at all, but at least they're now identified.

tradeitems.txt has been converted to the new format, and the version number
has been bumped.  Several combat item usage problems have been fixed, now
that it's possible to do so.


git-svn-id: svn://svn.code.sf.net/p/kolmafia/code@7094 b29ace70-8910-0410-8dcc-aa2fc6433167