Check for forumURL starting with, not equal to, newchatmessages.php or submitnewchat.php, to determine if a request is a chat request.
When the nuns claim Meat, log it in the session log:
The nuns take 2,167 Meat; 87,282 recovered, 12,718 left to recover.

Assume that anything less than a brigand's base meat drop - 800 - is from a
familiar action, regardless of meat drop percentage. It is possible for there to
be less than 800 Meat left to recover for the nuns. If so, meat drops of at
least that amount but less than 800 still count as going to the nuns.


git-svn-id: svn://svn.code.sf.net/p/kolmafia/code@7164 b29ace70-8910-0410-8dcc-aa2fc6433167