Emergency patch, part 2: allow multi-use of restoratives, deduct restoratives from inventory when used from the Skills page or quickskills menu. Remaining known issue: automated llama gong usage doesn't work. Untested: [use] links for restorative items that drop in combat.
When an ASH item object is created, always normalize the item name - even if
created from a known-good string.  Otherwise, items with aliases (such as El
Vibrato punchcards) can have strange problems: it was possible to create two
ASH items, referring to the same KoL item, that would act as distinct keys in
a map.


git-svn-id: svn://svn.code.sf.net/p/kolmafia/code@7379 b29ace70-8910-0410-8dcc-aa2fc6433167