DO NOT USE r7471!  All skill casts were being treated as maxcasts, ignoring your requested count.
Gory details: now that availableSkills is a sorted list, an instance of
UseSkillRequest has to be used as a key to see if the list contains that
skill.  UseSkillRequest.getInstance() seemed a reasonable way of getting such
a key, but it has side-effects - the target and count of the retrieved cached
instance are cleared.  Every cast request calls hasSkill() on itself to
verify that the request is valid, thus turning it into a request for a count
of zero on yourself.  Fixed by adding a getUnmodifiedInstance() method.


git-svn-id: svn://svn.code.sf.net/p/kolmafia/code@7473 b29ace70-8910-0410-8dcc-aa2fc6433167