Consider meat paste & stacks to be creatable for the purposes of item filtering - previously, "make dense meat sta" would match the staff, rather than giving an ambiguity error with the staff & stack.
Major speedup in the fuzzy string matcher - for each list of names that is
matched against, a corresponding list of hash values is precalculated.  The
majority of names that don't contain all of the distinct characters in the
search string can be skipped with a single integer operation, without looking
at the characters at all.  This saves about 5 million function calls during
KoLmafia startup.


git-svn-id: svn://svn.code.sf.net/p/kolmafia/code@7533 b29ace70-8910-0410-8dcc-aa2fc6433167